[Gold V] Title: 합분해, Time: 0 ms, Memory: 2336 KB -BaekjoonHub

+++ 백준/Gold/2225. 합분해/합분해.cc
@@ -2,7 +2,7 @@
 #define MOD 1000000000
 using namespace std;
 
-long long pas[402][201];
+int pas[402][201];
 
 int main() {
 	int N, K;
